Utilities to record installation activity
Adds utilites to track installation activity and produce an
"Installation Manifest".

Every filesystem path name which is modified during installation is
recorded along with metadata about the installation action and what
should be performed during an uninstall. The metadata is
extensible. The table can be formatted in a variety of ways, either as
a file which can be parsed (e.g. Installation Manifest), or as human
readable friendly summary information. The installation file can be
read later to perform an uninstall action.

Previously a less complete cleanup.dat file was produced which omitted
any information about files installed as part of a directory,
distinction beween symbolic links and files/directories, and what
should occur during an uninstall (e.g removal vs. preservation). The
utilities can detect the old file format utilize them it to preserve
backward compatibility.

Because the new format is extensible any future needs should be easily
accommodated.

Aside from a more complete and accurate manifest and user report there
is an additional benefit to this tracking information in terms of
developer debugging. I found this more detailed reporting invaluable
after modifying the installation script because it allowed me to see
if what I expected to happen was happening or if things which weren't
supposed to happen occurred. Formerly this was difficult information
to extract and has enhanced robustness, both during development and
during user install/uninstall.

This patch only adds the utilities, it does not invoke them.


git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1535 c9f7a03b-bd48-0410-a16d-cbbf54688b0b