Remove pointless no-op string interpolations
Constructs such as

"$variable"

when $variable is already a string are no-op waste of processor cycles
and confusing to read. Just use the variable.

Explanation:

Perl performs variable substitutions on all double quoted strings,
this is called string interpolation. To do this Perl scans the string
looking for anything that looks like a variable and substitutes it's
current value. But when the string consists of nothing other than a
variable (e.g. "$variable") the result is the same as the variable,
effectively it's just a no-op. I'm not sure if the interpreter is
smart enough to recognize this as a no-op and skip the interpolation,
irregardless there is no point in coding it this way.

It eludes me as to what the programmer thought they were accomplishing
when they wrote "$variable".


git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1542 c9f7a03b-bd48-0410-a16d-cbbf54688b0b