Update server.xml config file
This is mostly a merge of the tomcat 6 server.xml file with our
existing server.xml file from tomcat 5.

Merge in new comments.

remove org.apache.catalina.storeconfig.StoreConfigLifecycleListener
because it's not part of tomcat6

Parameterize the following based on our template engine:

sslOptions="[TOMCAT_SSL_OPTIONS]"
ssl2Ciphers="[TOMCAT_SSL2_CIPHERS]"
ssl3Ciphers="[TOMCAT_SSL3_CIPHERS]"
tls3Ciphers="[TOMCAT_TLS3_CIPHERS]"

Note: After the cipher parameterization was done it was discovered
that the other subsystems do not utilize ECC ciphers, it's not clear
if they should or not. We may need to paramterize the cipher list in
pkicreate or go back to hardcoding the cipher list in the xml file.


git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1561 c9f7a03b-bd48-0410-a16d-cbbf54688b0b