Bug 693815 - /var/log/tomcat6/catalina.out owned by pkiuser
Set the TOMCAT_LOG variable in the per instance tomcat
config file otherwise it defaults to the generic tomcat
log file. Note, we set up and configure our log file elsewhere
so the only issue was the initscript was setting the 
TOMCAT_USER ownership on TOMCAT_LOG, a file we otherwise do
not use or touch.


git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1954 c9f7a03b-bd48-0410-a16d-cbbf54688b0b