Change RecoveryRequest fields
When sending a passphrase in the recovery request, we need to wrap it
in a session key and store it in sessionWrappedPassphrase.  We also
then wrap the session key in transWrappedSessionKey.

The server needs to do PBE if the sessionWrappedPassphrase
is present, and symkey based encryption otherwise.

Also changed the DRM test to reflect these changes, and fixed some errors.