BZ 802396 - Change location of TOMCAT_LOG to match tomcat6 changes
Tomcat6 has changed the changed the location of the TOMCAT_LOG, and
it should no longer point to catalina.out.  This initially caused
dogtag to break because the code to chown TOMCAT_LOG to TOMCAT_USER
was removed.  Added code to spec file to fix existing instances.

Also fixed error in spec file.  Incorrect selinux patch was being
applied for f17.