Changes to use standard dbuser
We create a user that can be used to connect to the database using the
subsystem cert for client auth.  We identified this user, using the seeAlso
attribute and provided certmap rules to this effect.

For this user, we used to reuse the uid = user CA-hostname-port, which is already
created for inter-system communication.  But this is problematic if more than one
dbuser exists, as the directory server may bind as the incorrect user.  In any
replication topology, there must be only one dbuser using the subsystem cert.

To simplify things, we create a new user specifically for this purpose
(pkidbuser), and we remove the seeAlso attribute from the older dbusers.

A script is needed to convert existing dogtag 9 istances to use the new user,
and set the relevant acls.  This will be done in a separate commit.