Fixed error handling in RetrieveModificationsTask.
The RetrieveModificationsTask has been modified such that it can
recover from errors while still allowing graceful shutdown.

The task is scheduled to run once. When it's done it will schecule
another one depending on the situation. If the search is abandoned
or the connection is closed it will wait one minute before
reconnecting. If the system is being shutdown it will not
schedule any more task.

Ticket #365