I18n for ProfileList.template.
The messages in ProfileList.template in CA EE has been extracted
into a properties file which can be translated separately.

The original messages in the template have been marked as follows:
  <span class="message" name="...key...">...message...</span>

When the page is loaded into the browser, the original message will
be replaced with the translated messages.

Ticket #406