Simplified the configuration file using defaults.
Previously to create a subsystem the admin would have to copy the
entire default deployment configuration, which contains many
parameters, and then customize it. Now the deployment code has been
changed such that the default config file will be used to provide
the default values, so the admin will only need to provide the
non-default parameters, thus reducing the size of the file.

Sample configuration files are provided in /usr/share/pki/
deployment/config.

Ticket #399