Run restorecon on top-level log directory
We currently run a restorecon on the instance log directory, but not
on the top level log directory.  Restorecon is required for the top
level log directory since pkispawn creates it.  Without running a
restorecon, it gets the label of the parent directory (var_log_t)
instead of consulting the fcontext rule in the base policy and using
pki_var_log_t.

Ticket #431