Hardcode setting of resteasy-lib for instance
Tomcat in f17 expects the file under /etc/sysconfig/foo to be a
set of environment variables being set, and parses it that way.
We recently added some logic to source the global pki.conf file.
This works in f18, but breaks instance startup in f17.

While this works in f18, its an indication that we are using the
tomcat config file incorrectly.  Reverting to hardcoding resteasy lib.