Revert to using default config file for pkidestroy
Previously, we archived the default config file when an instance
was created, and used that file in running pkidestroy.  We plan
to replace this mechanism in favor of actually reading the instance's
config files.  For now, we return to using the standard default config
template, so that we can change it without breaking pkidestroy.