Added servlet to return 501 for rest operations for d9 instances
D9 instances run on tomcat6, which does not have support for the
autheticator and realm.  We are not supporting the REST operations
on D9 style instances.  They will need to be migrated.

The migration framework has been modified to process d9 or d10
style instances, and a migration script has been added to add the new
servlet to existing d9 instances.