Fix tests in pkispawn to use legacy URLs as fallback
When setting up clones or non-CA subsystems, pkispawn checks if
the security domain is accessible and if the user can log in.
These calls invoke REST URIs, which are not available on older
subsystems.  To support these subsystems, we need to attempt the
older legacy servlets if the REST APIs are not available.

Ticket #604