Fixed dependency issue on CMSRequest.
The CMSRequest is a server class but it's used by the ICommandQueue
that belongs in the base package. To fix the dependency issue the
CMSRequest has been refactored to implement a new interface
ICMSRequest in the base package. Some constants in CMSRequest have
also been moved into ICMSRequest. All code referencing CMSRequest
has been adjusted accordingly.