Storing authentication info in session.
The authenticator configuration has been modified to store the authentication
info in the session so it can be used by the servlets. An upgrade script has
been added to update the configuration in existing instances.

The SSLAuthenticatorWithFalback was modified to propagate the configuration
to the actual authenticator handling the request.