Provide enrollment template per profile
This adds an API call to get a template which can be used to generate an
enrollment request which can be passed into the REST API.  The template
is simply a CertRequest with the relevant inputs for that profile added in.

Per code review comments, have added the templates interface to
CertRequestResource instead.  This patch now includes /certrequests/profiles
and /certrequests/profiles/{id}.  In a subsequent patch, all calls in
ProfileResource will be restricted to admins and agents.