Catch all exceptions when checking for status.
python-requests now throws a ProxyError if the server is not yet up.
Previously only connect exceptions were seen.  To ensure that we are
not broken again when python-requests and the underlying libraries are
changed, we will catch and log all exceptions.  If the connection
ultimately fails, we will time out in any case.

Also fixed some new warnings from Pylint 1.0

Ticket 717