Modify TKS self tests and execution to use new shared secret names
The self tests and TokenServlet are modified to use the new shared secret
names.  A parameter has been added to allow legacy systems to continue running
as-is.  With a new system, the TKS self test will not fail on startup if
no shared secret keys are configured.  It will fail, however, if the keys are
configured, but the ComputeSessionKey operation fails.