Ticket 755 - Detect unescaped percent characters in deployment files
The deployment config files used by pkispawn support interpolation as
supplied by ConfigParser.  Interpolation uses the '%' character, which
means values that need to contain a '%' character need to be properly
escaped.

This patch detects errors with unescaped '%' characters and reports
a useful message bac kto the user who is running pkispawn.  This
patch also adds notes to the pkispawn and pki_default.cfg man pages
to explain that escaping of '%' characters is required.