Fixed database cleanup issues.
The configuration code has been modified not to remove the LDAP database
folder since it may not have access to it. It will also not continue
with the cleanup if the database is used by another subtree.

Manual removal of old entries in the subtree is redundant so the code
has been removed. The exception handling has been improved as well.