Debian: add init script functionality
The addtions in this patch will add start/stop/restart/status
functionality to operations, so that Debian systems can perform
these operations by calling these functions from an init script.

We also introduce a parameter in the configuration scripts that
can be used to determine if the system is a debian system.  This
parameter is used to specify a system V init script instead of
a systemd script on a debian system, when the configuration
scriptlets start and stop a system.

Also source apparently does not work by default in debian.  Used
dot (.) instead.