Improved TPS UI page loading.
Previously the TPS UI would reload the JS libraries and CSS files
every time a page is opened. The pages have been simplified such that
it will only contain the necessary elements of the page. A new Page
class has been added to the framework to define page initialization.

Ticket #654