Fixed problem refreshing table after add.
Previously the tables in TPS UI did not refresh properly after
adding a new entry. The code has been fixed to reset the collection
object to store the newly fetched entries and close the dialog after
the add operation is done. The success and error handlers for add
and save operations have been reorganized.

Ticket #654