Replaced CLI wrapper with Python.
The existing CLI wrapper script was written in Perl to call Java CLI.
It has been replaced with a Python script that can call either the
existing Java CLI or a not-yet-implemented Python CLI by specifying
a --client-type parameter. This will allow testing the Python client
library via CLI in the future.