Cannot unmarshall *Request objects to ResourceMessage object.
When using XML for communicating with the server, the JAXBContext is not
able to unmarshall KeyArchivalRequest,KeyRecoveryRequest, SymKeyGenerationRequest objects
as it does not recognize them to be subclasses of ResourceMessage. Adding a
@XmlSeeAlso attribute to the ResourceMessage class will let the JAXBContext to know that
the subclasses are binded to the super class and unmarshalls the *Request objects.

No problems were found when using JSON to communicate with the server.