Added remove button handler for TPS UI.
The Table class has been modified to handle the remove button.
When the button is clicked, it will get the list of items selected
for deletion. If there is nothing selected it will not do anything.
If there are some items selected, it will display the list in a
dialog box and ask the user to confirm the deletion. If the operation
is confirmed, the items will be deleted one by one. The table will
be refreshed after the process is done.

In some pages the buttons have been updated to reflect the
availability of the add/remove operations for the corresponding
REST resources.

Ticket #654