Added change token status dialog.
A new dialog has been added to change the token status. The status
can be changed by clicking the Status value in the tokens table.
Initially the status is Uninitialized. The status can be changed
according to the allowed status transitions defined in the CS.cfg.

The status and reason fields in TokenRecord is now translated into
a single status field in TokenData. This way the UI only needs to
handle a single status field.

A new field has also been added to the database for token type.
Some issues displaying and updating some token attributes have been
fixed as well.

Ticket #654