Replaced token dialog with token page.
The dialog used to edit token attributes has been replaced with a
details page since it will be required for breadcrumbs. A new HTML
template has been added for this page.

Changing token status now can be done both in token list page and
in token details page.

The EntryPage has been modified such that it requires the editable
fields to be specified for the add mode.

To improve the appearance, the input fields in all dialogs and pages
will now appear as read-only while the data is still loading.

Ticket #654