Refactored SystemConfigService (part 1).
This series of patches will incrementally refactor the
SystemConfigService to make it more manageable, which will
help development and troubleshooting, and also to resolve
potential dependency issues on subsystem-specific installation
procedures.

This patch converts the code that handles the certificate list
to use to Collection framework and to avoid duplicate parsing.
It also contains some code cleanup.

Ticket #890