Secure Channel final steps.
This patch allows the current secure channel functionality to work with
both tpsclient and esc. In order to get esc to work the following changes
were needed.

1. It turns out the server has been been forcing chunked encoding format upon the
outgoing data. Turns out that the system already knows how to do this so we were
getting double chunk size values and getting twice the amount of CRLF chars.

2. There was a minor error where I was not attempting to select the card manager
applet but the coolkey applet, which does not exist yet.