Added security domain functionality to python API
Currently the security domain python API just extracts the security
domain name from the json returned by the server.  This patch allows
it to extract and use all the information in the response.

This info is needed to determine the state of the security domain for
the IPA vault case.