Fix identities for security data storage, retrieval and generation
For the new security data storage and retrieval, and for symmetric
key generation, we need to store the identity of the agent that is
requesting and approving each operation, both in the ldap record
and in the audit logs.  (Tickets 806 and 807)

This patch also adds required logic to check that the owner of the
recovery request is the same agent that retrieves the key.  It also
adds missing audit log constants for symmmetric key generation so that
they will show up in the audit log.