Fixed NumberFormatException in key-request-find.
Previously if a key archival failed, the REST service would return
an invalid key URL, which would cause an exception when the CLI tried
to parse it. The service has been fixed to return a null URL which
can be detected to avoid parsing invalid value.

The Python library has been modified to handle missing key URL.

Ticket #1043