Add ability to create database as subtree of existing tree
This patch adds the ability to create a subsystem that uses
an existing subtree to create the internal basedn.  This is useful
for instance, for IPA which will use the original o=ipaca as the
top level DN for a KRA, which will be situated at o=ipadrm, o=ipaca.

The patch also allows such a system to be cloned, but not to setup the
replication agreements, on the assumption that the data is already being
replicated at the top-level DN or some higher level.

The patch also contains some minor cleanups - removing unused imports and
removal of an invalid reference in the python code.

Ticket 1051