Refactoring ProfileClient to remove the property fields.
Replaced the usage of python property feature with a dict
for attribute name conversion.

Fixed an issue caused to traversing the NOTYPES dict in
encoder.py to find the instance of an object. The traversal
causes an issue in the presence of subclassing.

Modified method attr_name_conversion to return a new dictionary with
modified attribute names rather than making changes to the object's
__dict__.