Fixed problem emptying a field in TPS UI.
Previously emptying a field in TPS UI could not be saved because
the change was not saved and sent to the server. The UI framework
now has been fixed to save and send the empty field to the server
such that the database can be updated properly.

Additional parameters have been added to the tps-token-mod command
to modify all editable fields.

Ticket #1085